Move scripts into src folder
