+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
   "include": [
     "src/**/*",
+    "scripts/*",
     "@types/*"
   ],
   "compilerOptions": {
